refactor(help): move help decorations into cli-framework
